Add /plot visit [player] command
This adds the ability to visit a player's claimed plot.

To accomplish this, I add a new function to `database` that can find a
plot's x, z coordinates by name, a new option to handle_plot_command,
and new entries to the completion tree.

I tested this to work for multiple players, and it does not break if
someone has multiple plots, though you will always only go to one.